projects/haikubot/requirements.txt
python-twitter
requests
nltk
